feat(wip): added new zustand store with multiple message types
This is working up to being able to send a message. After that while
chunks are streaming, they are not being updated on the UI. Will fix
this next. Broke an infinite rerender cycle where zustand store update
was causing infinte rerenders. Somehow fixed it.

In this setup:
- On receiving ACK a new streamMessage is constructed
- We have HumanMessage and intelligenceMessage which both extend
  BaseMessage. The IntelligenceMessage has streamId and RequestId that
  gets sent in the ACK
- replaced humanMessages, generativeMessages and so forth with the
  message store ones.

Bugs: The streaming chunks are not showing in the UI.